Keep user-defined width and height.
This results in an aspect ratio that might not be the same as the Storage Aspect Ration (SAR), however, in some cases, the video might have a different Display Aspect Ratio (DAR) than SAR, so this gives the user a chance to get the ratio right. Besides, if the user sets the dimensions they should get full control rather than one of their dimensions getting overriden by smart logic.