Always have only single 404 route as the last one
In `boot/index`, add a 404 route as the last route before starting
`page` library. If code-splitting is disabled and no route is matched,
this 404 handler will be called. If code-splitting is enabled, in the
async loader of a section, remove the current (last) 404 route, then
load the section's routes and finally register a new 404 route. This
way, just a single 404 route will be present in the app at a time.